test(intellij): relocate #3624 addNotify test to ShaftAssistantPanelTest
The #3624 test was added to ShaftPanelSetupTest.java because
ShaftAssistantPanelTest.java (the #3621 fix's natural home, with the
same reflection-injection pattern for connectionState/transcript/
lafConnectionDisposable) didn't exist yet on this branch. Now that
Phase 1 is merged in, move it there for consistency; drop the
resulting unused ShaftMcpConnectionState import from
ShaftPanelSetupTest.java.
